+++ app/Core/Http/Traits/RequestTrait.php
@@ -98,7 +98,7 @@ trait RequestTrait
     public function withRequestTarget($requestTarget)
     {
         if (preg_match('#\s#', $requestTarget)) {
-            throw new InvalidArgumentException('Invalid request target provided. Must be a string without whitespace');
+            InvalidArgumentException::alertMessage(400, 'Invalid request target provided. Must be a string without whitespace');
         }
 
         $clone = clone $this;
@@ -146,7 +146,7 @@ trait RequestTrait
         }
 
         if (! is_string($method)) {
-            throw new InvalidArgumentException(
+            InvalidArgumentException::alertMessage(400, 
                 'Invalid HTTP method. Must be a string, received ' .
                 (is_object($method) ? get_class($method) : gettype($method))
             );
@@ -155,7 +155,7 @@ trait RequestTrait
         $method = strtoupper($method);
 
         if (! in_array($method, self::$validMethods, true)) {
-            throw new InvalidArgumentException(
+            InvalidArgumentException::alertMessage(400, 
                 'Invalid HTTP method. Must be ' .
                 implode(', ', self::$validMethods)
             );
@@ -181,7 +181,7 @@ trait RequestTrait
         } elseif ($uri === null) {
             $this->uri = new Uri;
         } else {
-            throw new InvalidArgumentException(
+            InvalidArgumentException::alertMessage(400, 
                 'Invalid URI provided. Must be null, a string, ' .
                 'or a Psr\Http\Message\UriInterface instance'
             );